+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### v4.51.0 Issues
 
+- `[Circlepager]` Fixed a bug where circle buttons doesn't work on smaller viewport and first initialization of the page. ([#4966](https://github.com/infor-design/enterprise/issues/4966))
 - `[General]` The master branch is now called main. Also cleaned up some language in the repo known to be less inclusive. ([#5027](https://github.com/infor-design/enterprise/issues/5027))
 
 ## v4.50.0